pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cs"
version = "0.0.1"
description = "Comp Stats project"
requires-python = ">=3.8"
dependencies = [
    "tensorboard==2.14.0",
    "lightning==2.1.2",
    "scikit-learn==1.3.2",
    "matplotlib",
    "natsort==8.4.0",
    "hydra-core==1.3.2",
    "pandas",
    "scipy==1.11.4",
    "umap-learn==0.5.5"
]

[tool.setuptools.packages.find]
include = ["cs"]

[project.scripts]
csrun = "cs.main:main"
csimput = "cs.imputation:main"
cstest = "cs.test_model:main"

[project.optional-dependencies]
cuda = ["torch==2.1.2+cu121"]
cpu = ["torch==2.1.2"]